+++ src/argument/argument_impl.rs
@@ -35,12 +35,12 @@ where
     }
 }
 
-impl<A, O, C> Execute for DefaultExecutor<A, C, O>
+impl<A, O, C, U> Execute<U> for DefaultExecutor<A, C, O>
 where
     A: CommandArgument<O>,
-    C: TaskLogic<O>,
+    C: TaskLogic<O, Output = U>,
 {
-    fn execute<'a>(&self, input: &'a str) -> IResult<&'a str, bool, CommandError<'a>> {
+    fn execute<'a>(&self, input: &'a str) -> IResult<&'a str, U, CommandError<'a>> {
         let (input, result) = self.argument.parse(input)?;
         match self.task.run(result) {
             Err(e) => Err(nom::Err::Failure(CommandError::from_external_error(input, ErrorKind::MapRes, e))),
@@ -49,13 +49,13 @@ where
     }
 }
 
-impl<A, O, C, T> Propagate<T> for DefaultExecutor<A, C, O>
+impl<A, O, C, T, U> Propagate<T, U> for DefaultExecutor<A, C, O>
 where
     T: Copy,
     A: CommandArgument<O>,
-    C: TaskLogic<(T, O)>,
+    C: TaskLogic<(T, O), Output = U>,
 {
-    fn propagate<'a>(&self, input: &'a str, data: T) -> IResult<&'a str, bool, CommandError<'a>> {
+    fn propagate<'a>(&self, input: &'a str, data: T) -> IResult<&'a str, U, CommandError<'a>> {
         let (input, result) = self.argument.parse(input)?;
         match self.task.run((data, result)) {
             Err(e) => Err(nom::Err::Failure(CommandError::from_external_error(input, ErrorKind::MapRes, e))),
@@ -64,22 +64,24 @@ where
     }
 }
 
-impl<E, F, O> TaskLogic<O> for F
+impl<E, F, O, U> TaskLogic<O> for F
 where
-    F: Fn(O) -> Result<bool, E>,
+    F: Fn(O) -> Result<U, E>,
     E: Into<anyhow::Error>,
 {
     type Error = E;
+    type Output = U;
 
-    fn run(&self, args: O) -> Result<bool, E> { self(args) }
+    fn run(&self, args: O) -> Result<U, E> { self(args) }
 }
 
-impl<E, F> TaskLogicNoArgs for F
+impl<E, F, U> TaskLogicNoArgs for F
 where
-    F: Fn() -> Result<bool, E>,
+    F: Fn() -> Result<U, E>,
     E: Into<anyhow::Error>,
 {
     type Error = E;
+    type Output = U;
 
-    fn run(&self) -> Result<bool, Self::Error> { self() }
+    fn run(&self) -> Result<U, Self::Error> { self() }
 }