+++ src/parsers/then.rs
@@ -14,23 +14,23 @@ pub struct ThenWrapper<E1, E2> {
     pub(crate) second: E2,
 }
 
-impl<E1, E2> Execute for ThenWrapper<E1, E2>
+impl<E1, E2, U> Execute<U> for ThenWrapper<E1, E2>
 where
-    E1: Execute,
-    E2: Execute,
+    E1: Execute<U>,
+    E2: Execute<U>,
 {
-    fn execute<'a>(&self, input: &'a str) -> IResult<&'a str, bool, CommandError<'a>> {
+    fn execute<'a>(&self, input: &'a str) -> IResult<&'a str, U, CommandError<'a>> {
         alt((|i| self.first.execute(i), |i| self.second.execute(i)))(input)
     }
 }
 
-impl<E1, E2, T> Propagate<T> for ThenWrapper<E1, E2>
+impl<E1, E2, T, U> Propagate<T, U> for ThenWrapper<E1, E2>
 where
     T: Copy,
-    E1: Propagate<T>,
-    E2: Propagate<T>,
+    E1: Propagate<T, U>,
+    E2: Propagate<T, U>,
 {
-    fn propagate<'a>(&self, input: &'a str, data: T) -> IResult<&'a str, bool, CommandError<'a>> {
+    fn propagate<'a>(&self, input: &'a str, data: T) -> IResult<&'a str, U, CommandError<'a>> {
         alt((|i| self.first.propagate(i, data), |i| self.second.propagate(i, data)))(input)
     }
 }